Sheer laziness, and in some part the excitement from the hostage release, has kept me from making a commit the last few days.
This last commit features a working scrollbar. It needs some polish, but when all is said and done, it does what's required of it.

Another milestone, is having the code work on two compilers, passing all errors and warning, including those from static analyzers. Wow, what a chore, that was...

What's next in store for the project?